requirements.txt
pandas~=1.1.3
python-dateutil~=2.8.1
numpy~=1.17.4
scipy~=1.5.3
streamlit~=0.69.2